# backend/docs/openapi.json
{"openapi": "3.0.2", "info": {"title": "datastation", "version": "0.1.0"}, "paths": {"/api/v1/login/access-token": {"post": {"tags": ["login"], "summary": "Login Access Token", "description": "OAuth2 compatible token login, get an access token for future requests", "operationId": "login_access_token_api_v1_login_access_token_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_access_token_api_v1_login_access_token_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/login/test-token": {"post": {"tags": ["login"], "summary": "Test Token", "description": "Test access token", "operationId": "test_token_api_v1_login_test_token_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/reset-password/": {"post": {"tags": ["login"], "summary": "Reset Password", "description": "Reset password", "operationId": "reset_password_api_v1_reset_password__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_reset_password_api_v1_reset_password__post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/users/": {"get": {"tags": ["users"], "summary": "Read Users", "description": "Retrieve users.", "operationId": "read_users_api_v1_users__get", "parameters": [{"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 100}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Read Users Api V1 Users  Get", "type": "array", "items": {"$ref": "#/components/schemas/User"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["users"], "summary": "Create User", "description": "Create new user.", "operationId": "create_user_api_v1_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/users/me": {"get": {"tags": ["users"], "summary": "Read User Me", "description": "Get current user.", "operationId": "read_user_me_api_v1_users_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["users"], "summary": "Update User Me", "description": "Update own user.", "operationId": "update_user_me_api_v1_users_me_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_update_user_me_api_v1_users_me_put"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/v1/users/open": {"post": {"tags": ["users"], "summary": "Create User Open", "description": "Create new user without the need to be logged in.", "operationId": "create_user_open_api_v1_users_open_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_create_user_open_api_v1_users_open_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/users/{user_id}": {"get": {"tags": ["users"], "summary": "Read User By Id", "description": "Get a specific user by id.", "operationId": "read_user_by_id_api_v1_users__user_id__get", "parameters": [{"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["users"], "summary": "Update User", "description": "Update a user.", "operationId": "update_user_api_v1_users__user_id__put", "parameters": [{"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}}, "components": {"schemas": {"Body_create_user_open_api_v1_users_open_post": {"title": "Body_create_user_open_api_v1_users_open_post", "required": ["password", "email"], "type": "object", "properties": {"password": {"title": "Password", "type": "string"}, "email": {"title": "Email", "type": "string", "format": "email"}, "full_name": {"title": "Full Name", "type": "string"}}}, "Body_login_access_token_api_v1_login_access_token_post": {"title": "Body_login_access_token_api_v1_login_access_token_post", "required": ["username", "password"], "type": "object", "properties": {"grant_type": {"title": "Grant Type", "pattern": "password", "type": "string"}, "username": {"title": "Username", "type": "string"}, "password": {"title": "Password", "type": "string"}, "scope": {"title": "Scope", "type": "string", "default": ""}, "client_id": {"title": "Client Id", "type": "string"}, "client_secret": {"title": "Client Secret", "type": "string"}}}, "Body_reset_password_api_v1_reset_password__post": {"title": "Body_reset_password_api_v1_reset_password__post", "required": ["token", "new_password"], "type": "object", "properties": {"token": {"title": "Token", "type": "string"}, "new_password": {"title": "New Password", "type": "string"}}}, "Body_update_user_me_api_v1_users_me_put": {"title": "Body_update_user_me_api_v1_users_me_put", "type": "object", "properties": {"password": {"title": "Password", "type": "string"}, "full_name": {"title": "Full Name", "type": "string"}, "email": {"title": "Email", "type": "string", "format": "email"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Message": {"title": "Message", "required": ["msg"], "type": "object", "properties": {"msg": {"title": "Msg", "type": "string"}}}, "Token": {"title": "Token", "required": ["access_token", "token_type"], "type": "object", "properties": {"access_token": {"title": "Access Token", "type": "string"}, "token_type": {"title": "Token Type", "type": "string"}}}, "User": {"title": "User", "required": ["created_at"], "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "full_name": {"title": "Full Name", "type": "string"}}}, "UserCreate": {"title": "UserCreate", "required": ["email", "password"], "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "full_name": {"title": "Full Name", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "UserUpdate": {"title": "UserUpdate", "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}, "is_active": {"title": "Is Active", "type": "boolean", "default": true}, "full_name": {"title": "Full Name", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/api/v1/login/access-token"}}}}}}